Elixir.Diode v0.0.1 Chain

Types

t()

t() :: %Chain{
  by_hash: %{required(binary()) => Chain.Block.t()} | nil,
  final: Chain.Block.t(),
  peak: Chain.Block.t(),
  states: Map.t(),
  window: %{required(binary()) => integer()}
}

Functions

add_block(block, relay \\ true)

averageTransactionGas()

averageTransactionGas() :: 200_000

block(n)

block(number()) :: Chain.Block.t() | nil

block_by_hash(hash)

block_by_hash(any()) :: Chain.Block.t() | nil

blockchainDelta()

blockchainDelta() :: 5

blocks()

blocks() :: Enumerable.t()

blocks(block)

blocks(any()) :: Enumerable.t()

blocktimeGoal()

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

epoch()

epoch_length()

final_block()

final_block() :: Chain.Block.t()

gasLimit()

Gaslimit for block validation and estimation

gas_price()

GasPrice for block validation and estimation

generate_blockquick_window(final_block)

genesis_hash()

init(_)

init(any()) :: {:ok, Chain.t()}

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

init_arg is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state}, except that it also sets a timeout. See the "Timeouts" section in the module documentation for more information.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

load_file(filename, default \\ nil)

peak()

peak() :: integer()

peak_block()

peak_block() :: Chain.Block.t()

peak_state()

peak_state() :: Chain.State.t()

pre_genesis_hash()

print_transactions(block)

reseed()

reseed is a debugging function

reset_state()

Function for unit tests, resets state to genesis state

saver_loop()

saver_loop() :: no_return()

set_final_block(block)

set_state(state)

Function for unit tests, replaces the current state

start_link(opts)

start_link(any()) :: :ignore | {:error, any()} | {:ok, pid()}

state()

state(n)

state(number()) :: Chain.State.t()

store_file(filename, term, overwrite \\ false)

sync()