Elixir.Diode v0.0.1 Chain.BlockCache

Functions

cache(block)

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

coinbase(block)

See Chain.Block.coinbase/1.

create(parent, transactions, miner, time, trace? \\ false)

See Chain.Block.create/5.

create_cache(block)

difficulty(block)

encode_transactions(transactions)

See Chain.Block.encode_transactions/1.

epoch(block)

extraData(block)

See Chain.Block.extraData/1.

gasLimit(block)

See Chain.Block.gasLimit/1.

gasUsed(block)

See Chain.Block.gasUsed/1.

gas_price(block)

See Chain.Block.gas_price/1.

has_state?(block)

See Chain.Block.has_state?/1.

hash(block)

See Chain.Block.hash/1.

hash_in_target?(block, hash)

See Chain.Block.hash_in_target?/2.

hash_target(block)

See Chain.Block.hash_target/1.

hash_valid?(block)

See Chain.Block.hash_valid?/1.

header(block)

See Chain.Block.header/1.

increment_nonce(block)

See Chain.Block.increment_nonce/1.

init(_)

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

init_arg is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state}, except that it also sets a timeout. See the "Timeouts" section in the module documentation for more information.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

logs(block)

See Chain.Block.logs/1.

logsBloom(block)

See Chain.Block.logsBloom/1.

miner(block)

See Chain.Block.miner/1.

nonce(block)

See Chain.Block.nonce/1.

number(block)

See Chain.Block.number/1.

parent(block)

See Chain.Block.parent/1.

parent_hash(block)

See Chain.Block.parent_hash/1.

receipts(block)

See Chain.Block.receipts/1.

receiptsRoot(block)

See Chain.Block.receiptsRoot/1.

reset()

save()

sign(block, priv)

See Chain.Block.sign/2.

simulate(block, trace? \\ false)

See Chain.Block.simulate/2.

size(block)

See Chain.Block.size/1.

start_link(arg)

state(block)

See Chain.Block.state/1.

state_hash(block)

See Chain.Block.state_hash/1.

timestamp(block)

See Chain.Block.timestamp/1.

totalDifficulty(block)

transactionGas(block, transaction)

See Chain.Block.transactionGas/2.

transactionIndex(block, transaction)

See Chain.Block.transactionIndex/2.

transactionOut(block, transaction)

See Chain.Block.transactionOut/2.

transactionStatus(block, transaction)

See Chain.Block.transactionStatus/2.

transactions(block)

See Chain.Block.transactions/1.

txhash(block)

See Chain.Block.txhash/1.

valid?(block)

See Chain.Block.valid?/1.

validate(block)

See Chain.Block.validate/1.