Elixir.Diode v0.0.1 Debounce

Functions

apply(key, fun, timeout \\ 5000)

apply(term(), (() -> any()), non_neg_integer()) :: :ok

apply executes the function after the specified timeout t0 + timeout,

when apply is called multiple times it does not affect the point
in time when the next call is happening (t0 + timeout) but updates the fun

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

delay(key, fun, timeout \\ 5000)

delay(term(), (() -> any()), non_neg_integer()) :: :ok

delay executes the function after the specified timeout t0 + timeout,

when delay is called multipe times the timeout is reset based on the
most recent call (t1 + timeout, t2 + timeout) etc... the fun is also updated

immediate(key, fun, timeout \\ 5000)

immediate(term(), (() -> any()), non_neg_integer()) :: :ok

immediate executes the function immediately but blocks any further call

under the same key for the given timeout.

init(arg)

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

init_arg is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state}, except that it also sets a timeout. See the "Timeouts" section in the module documentation for more information.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

start_link(args)

start_link(any()) :: :ignore | {:error, any()} | {:ok, pid()}