Elixir.Diode v0.0.1 KBuckets

Provides 256 bits k-buckets for hashes

Types

item()

item() :: %KBuckets.Item{
  last_seen: integer(),
  node_id: Wallet.t(),
  object: Object.Server.server() | :self,
  retries: any()
}

item_id()

item_id() :: <<_::256>>

kbuckets()

kbuckets() :: {:kbucket, any(), tree() | nil}

node_id()

node_id() :: Wallet.t()

tree()

tree() :: {:leaf, any(), map()} | {:node, any(), tree(), tree()}

Functions

bucket_count(arg)

bucket_count(kbuckets()) :: pos_integer()

delete_item(arg, item)

delete_item(kbuckets(), item() | item_id()) :: kbuckets()

distance(a, b)

Calculates the linear distance on a geometric ring from 0 to 2^256 any distance. Because it's a ring the maximal distance is two points being opposite of each other. In that case the distance is 2^255.

2^256 = 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff 2^255 = 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff

hash(w)

insert_item(arg, item)

insert_item(kbuckets(), item() | item_id()) :: kbuckets()

insert_items(arg, items)

insert_items(kbuckets(), [item()]) :: kbuckets()

is_self(bucket)

item(kb, wallet)

k()

k() :: integer()

key(arg1)

member?(list, item)

nearer_n(kbuckets, item, n)

nearer_n(kbuckets() | [item()], item() | item_id(), pos_integer()) :: [item()]

nearer_n finds the n nodes nearer or equal to the current node to the provided item.

nearest_n(list, item, n)

nearest_n(kbuckets() | [item()], item() | item_id(), pos_integer()) :: [item()]

nearest_n finds the n nodes nearer or equal to the current node to the provided item.

new(self_id)

new(node_id()) :: kbuckets()

object(item)

self(arg)

self(kbuckets()) :: item()

size(arg)

size(kbuckets()) :: non_neg_integer()

to_list(arg)

to_list(kbuckets()) :: [item()]

to_uri(item)

to_uri(item()) :: binary()

unique(list)

update_item(arg, item)

update_item(kbuckets(), item() | item_id()) :: kbuckets()