Elixir.Diode v0.0.1 Kademlia

Types

t()

t() :: %Kademlia{cache: Lru.t(), network: KBuckets.t(), tasks: Map.t()}

Functions

append(key, value, store_self \\ false)

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

find_node(key)

find_node(any()) :: nil | KBuckets.Item.t()

find_node_lookup(key)

find_node_lookup(any()) :: [KBuckets.item()]

find_nodes(key)

find_nodes(any()) :: [KBuckets.Item.t()]

find_value(key)

find_value(binary()) :: any()

find_value() is different from store() in that it might return an earlier result

handle_continue(atom, state)

Invoked to handle continue instructions.

It is useful for performing work after initialization or for splitting the work in a callback in multiple steps, updating the process state along the way.

Return values are the same as c:handle_cast/2.

This callback is optional. If one is not implemented, the server will fail if a continue instruction is used.

This callback is only supported on Erlang/OTP 21+.

Callback implementation for GenServer.handle_continue/2.

init(atom)

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

init_arg is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state}, except that it also sets a timeout. See the "Timeouts" section in the module documentation for more information.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

network()

oid_multiply(oid, factor)

oid_multiply(<<_::256>>, integer()) :: [any()]

ping(node_id)

publish(object)

reset()

Method used for testing

rpc(nodes, call)

start_link(opts)

store(key, value)

store(binary(), any()) :: any()

store() stores the given key-value pair in the @replication_factor nodes that are closest to the key