Elixir.Diode v0.0.1 Object behaviour

All objects are made of tuples {:type, value1, value2, ..., valueN, signature} The number of values are different but the last signature is a signature is always is the signature of BertExt.encode!([value1, value2, ..., valueN])) Also the signatures public key is always equal to the key

Types

key()

key() :: <<_::160>>

Functions

decode!(bin)

decode_list!(list)

encode!(record)

encode_list!(record)

key(record)

key(tuple()) :: binary()

ticket_id(record)

Callbacks

key(tuple)

key(tuple()) :: key()