Elixir.Diode v0.0.1 Wallet

Might be merged with Id.ex, for now just starting with clear Ethereum triple needed for wallets: 1) Private Key 2) Public Key 3) Address A wallet can be instantiated from any of those, but accessors for higher level ids fail

Types

address()

address() :: <<_::160>>

compressed_public_key()

compressed_public_key() :: <<_::264>>

private_key()

private_key() :: <<_::256>>

t()

t() ::
  {:wallet, privkey :: nil | <<_::256>>, pubkey :: nil | <<_::264>>,
   address :: <<_::160>>}

Functions

address(arg)

address(Wallet.t()) :: {:error, nil} | {:ok, <<_::160>>}

address!(wallet)

address!(Wallet.t()) :: <<_::160>>

base16(wallet)

equal?(a, b)

from_address(address)

from_address(<<_::160>> | integer()) :: Wallet.t()

from_privkey(privkey)

from_privkey(<<_::256>> | integer()) :: Wallet.t()

from_pubkey(pubkey)

from_pubkey(<<_::264>>) :: Wallet.t()

new()

new() :: Wallet.t()

nick(wallet)

printable(wallet)

privkey(arg)

privkey!(wallet)

pubkey(arg)

pubkey!(wallet)

pubkey_long(arg)

pubkey_long!(wallet)

wallet(args \\ [])
(macro)

wallet(record, args)
(macro)

words(wallet)